\name{splinesurv}
\alias{splinesurv}
\alias{splinesurv.formula}
\alias{splinesurv.agdata}
\alias{splinesurv.data.frame}
\alias{print.splinesurv}
\alias{splinesurv.object}
\title{Bayesian survival analysis with B-spline baseline hazard and random effects density}
\description{Allows the fitting of proportional hazards survival models
    to possibly clustered data using Bayesian methods. The function follows
    a MCMC method to sample from the posterior distribution of the regression
    parameters, frailties, and parameters specifying the hazard and frailty
    distribution.
    
    The baseline hazard and random effects desity are modeled as penalized B-splines,
    with the penalty depending on either the integrated squared second derivative
    of the curve, or the sum of squared second differences in the spline parameters.
}
\synopsis{\method{splinesurv}{formula}(formula, data = parent.frame(), ...)}
\usage{
\method{splinesurv}{formula}(formula, data = parent.frame(), verbose = 3,
           control = NULL, spline = NULL, initial = NULL, coda = FALSE,...)
}
\arguments{
    \item{formula}{a formula object, similar to \code{\link{coxph}}. The response to the
        left of the \code{~} should be a survival object generated by \code{\link{Surv}}. The
        right side may contain a \code{cluster(x)} term if \code{x} is
        the variable that indicates cluster membership.
    }
    \item{data}{a \code{data.frame} with columns corresponding to the terms 
        in the formula.}
    \item{verbose}{an integer from 0 to 5 that determines the quantity of output
        printed to the screen. Setting \code{verbose=0} is completely silent.}
    \item{control}{a list containing control parameters for the MCMC and optimization,
        with the following optional components. For any component
        that is not set, the default is used.
        \describe{
            \item{\code{burnin}}{an integer giving the number of iterations discarded
                as burn-in. Default is \code{burnin=500}.}
            \item{\code{maxiter}}{an integer giving the total number of MCMC iterations,
                must be greater than \code{burnin}. Default is \code{maxiter=1000}.}
            \item{\code{cal.gamma}}{logical, determines whether to adaptively calibrate
                the tuning parameters 
                \eqn{(\gamma_U,\gamma_\beta,\gamma_\lambda,\gamma_u)}{(gamma.U, gamma.b, gamma.l, gamma.u)}
                during the burn-in period to achieve an acceptance rate of 0.25. 
                Defaults to \code{TRUE}.}
            \item{\code{calint}}{an integer giving number of iterations between calibration
                of the tuning parameters, if \code{cal.gamma=TRUE}. 
                Default is \code{calint=100}.}
            \item{\code{gamma}}{a vector of length 4, containing initial values for the
                tuning parameters \eqn{(\gamma_U,\gamma_\beta,\gamma_\lambda,\gamma_u)}{(gamma.U, gamma.b, gamma.l, gamma.u)}
                in that order. If \code{cal.gamma=FALSE}, they remain unchanged. Defaults to
                \code{gamma=rep(1,4)}}
            \item{\code{alpha}}{a vector of length 6, giving the two hyperparameters for
                the inverse-gamma priors of \eqn{\sigma^2_\beta,\sigma^2_\lambda,\sigma^2_u}{sigma2.b, sigma2.l, sigma2.u}
                in that order. Defaults to \code{alpha=rep(0.01,6)}.}
            \item{\code{M}}{a large number determining the weight of the penalty to ensure that
                the mean of the frailty density is 1. Defaults to \code{M=1000}.}
        }
    }
    \item{spline}{a list containing parameters defining the B-splines for the
        baseline hazard and frailty distribution, with the following optional
        components. For any component that is not set, the default is used.
        \describe{
            \item{\code{ord.haz}}{an integer greater than 1, giving the order of 
                the spline defining the baseline hazard. Defaults to
                \code{ord.haz=4}, corresponding to cubic B-splines.}
            \item{\code{nknots.haz}}{an integer giving the number of interior
                knots used in the baseline hazard spline. If \code{NULL}, the
                number of knots is chosen automatically. Defaults to \code{NULL}.}
            \item{\code{knotspacing.haz}}{string that determines the way that knots are
                automatically chosen, if applicable. Possible values are \code{"quantile"}
                to place knots at quantiles of observed event times, or \code{"equal"} to
                equally space knots over the range of observed times. Defaults to \code{"quantile"}.}
            \item{\code{knots.haz}}{a vector of length \code{knots.haz+2*ord.haz-2} giving the 
                positions of all the knots used in the baseline hazard spline (including
                boundary knots). If \code{NULL}, knots are chosen automatically. 
                Defaults to \code{NULL}.}
            \item{\code{ord.frail}}{analogous to \code{ord.haz}, but for the frailty density
                spline.}
            \item{\code{nknots.frail}}{analogous to \code{nknots.haz}.}
            \item{\code{knotspacing.frail}}{analogous to \code{knotspacing.haz}, except that \code{"quantile"}
                is not allowed.}
            \item{\code{knots.frail}}{analogous to \code{knots.haz}.}
            \item{\code{penalty.haz}}{a string giving the type of penalty to be used for
                the baseline hazard splines. Possible values are \code{"2diff"}, for
                a penalty on the second differences, or \code{"2deriv"} for a penalty on
                the integrated squared second derivative. Defaults to \code{"2deriv"}.}
            \item{\code{penalty.frail}}{a string giving the type of penalty to be used for
                the frailty density splines. Analogous to \code{penalty.haz}, but defaults
                to \code{"2diff"}.}
        }
    }
    \item{initial}{a list containing initial values for the chain. Not implemented and
        currently ignored.}
    \item{coda}{a logical variable indicating whether the \code{coda} package should be used to return
        the parameter history as \code{mcmc} objects. Defaults to \code{FALSE}.}
    \item{...}{additional parameters (currently ignored).}
}
\value{
    An object of class \code{splinesurv}, with the following components
    \item{call}{the original call to the model-fitting function}
    \item{history}{a list containing the parameter history of the 
        MCMC iterations, with the following components:
        \describe{
            \item{\code{frailty}}{matrix of frailty estimates at each iteration.}
            \item{\code{coefficients}}{matrix of regression parameter estimates at each iteration.}
            \item{\code{splinepar.haz}}{matrix of parameters defining the baseline hazard spline.}
            \item{\code{splinepar.frail}}{matrix of parameters defining the frailty density spline.}
            \item{\code{sigma2}}{matrix of dispersion parameters for 
            the regression parameters, hazard spline, and frailty density spline in that order.}
            \item{\code{acc}}{binary vectors of Metropolis-Hastings acceptance
            indicators for the frailties, the regression parameters, hazard spline, and frailty 
            density spline respectively.}
        }
        If in the input, \code{coda=TRUE}, then the returned objects are of class \code{mcmc} instead of \code{matrix}.
    }
    \item{spline}{a list analogous to \code{spline} in the input, describing the final knot
        placement for the baseline hazard and frailty density splines. If knots were pre-determined,
        this is the same as the input, but if they were chosen automatically, this contains the 
        selected knots.}
    \item{control}{a list analogous to \code{control} in the input, containing the control
        parameters used in the procedure, even those not set explicitly in the call.}
}
\references{E. Sharef, D. Ruppert and R. Strawderman. \dQuote{A Bayesian approach 
to clustered survival analysis with nonparametric baseline hazard and frailty 
distributions}, \emph{in preparation}.
}
\seealso{
\code{\link{summary.splinesurv}}, \code{\link{plot.splinesurv}}, \code{\link{coxph}}
}
\examples{
    ## Generate a small survival data set:
    s <- sim.sample(m = 10, Ji = rep(10, 10))
    agdata <- s$agdata
    
    ## Run a (very) short MCMC chain
    fit <- splinesurv(Surv(time, delta) ~ Z + cluster(i), data = agdata, control = list(maxiter = 100, burnin = 50))
    
    ## Run another chain,  with linear B-splines,  
    # and a different penalty and knotspacing for the hazard
    fit2 <- splinesurv(Surv(time, delta) ~ Z + cluster(i), data = agdata, control = list(maxiter = 100, burnin = 50), spline = list(ord.frail = 2, ord.haz = 2, knotspacing.haz = "equal"))

    ## View summaries and plots of the fits
    summary(fit)
    plot(fit, "all")
    summary(fit2)
    plot(fit2, "all")
}
\keyword{survival}